revert: refactor of run
