+++ lib/monitor/run.js
@@ -14,29 +14,50 @@ var restart = null;
 var psTree = require('pstree.remy');
 var path = require('path');
 var signals = require('./signals');
-const getSpawnArgs = require('./spawn-args');
 
 function run(options) {
+  var cmd = config.command.raw;
+
+  var runCmd = !options.runOnChangeOnly || config.lastStarted !== 0;
+  if (runCmd) {
+    utils.log.status('starting `' + config.command.string + '`');
+  }
+
   /*jshint validthis:true*/
   restart = run.bind(this, options);
   run.restart = restart;
 
   config.lastStarted = Date.now();
 
-  const {
-    args,
-    cmd,
-    executable,
-    runCmd,
-    sh,
-    shFlag,
-    spawnArgs,
-    stdio,
-  } = getSpawnArgs({ options, config });
+  var stdio = ['pipe', 'pipe', 'pipe'];
 
-  if (runCmd) {
-    utils.log.status('starting `' + config.command.string + '`');
+  if (config.options.stdout) {
+    stdio = ['pipe', process.stdout, process.stderr];
   }
+
+  if (config.options.stdin === false) {
+    stdio = [process.stdin, process.stdout, process.stderr];
+  }
+
+  var sh = 'sh';
+  var shFlag = '-c';
+
+
+  const spawnOptions = {
+    env: utils.merge(options.execOptions.env, process.env),
+    stdio: stdio,
+  }
+
+  if (utils.isWindows) {
+    // taken from npm's cli: https://git.io/vNFD4
+    sh = process.env.comspec || 'cmd';
+    shFlag = '/d /s /c';
+    spawnOptions.windowsVerbatimArguments = true;
+  }
+
+  var executable = cmd.executable;
+  var args = runCmd ? utils.stringify(executable, cmd.args) : ':';
+  var spawnArgs = [sh, [shFlag, args], spawnOptions];
 
   const firstArg = cmd.args[0] || '';
